requirements.txt
Django==4.2
asgiref==3.6.0
djangorestframework==3.14.0
pytz==2023.3
setuptools==67.6.1
sqlparse==0.4.4
tzdata==2023.3
wheel==0.40.0